Create and edit assessments
Repairs can now be assessed using the assess action.

A job type can be selected, auto filling time, materials and notes.

Time can be changed, and time cost will be auto calculated based on hourly rate in settings. This can be manually overrided.

Materials can be added and removed, and quantities can be changed. Costs will be auto calculated based on materials settings table, but they can be manually changed.

Notes can be edited. Job type is permanent after initial assessment.

Overwriting an assessment changes the existing assessment. This is the only option for older assessments.

Updating the existing assessment creates a new assessment with the new data, which can be switched to and from using a dropdown. This is to keep a history of assessments.